+++ packages/tanstack_query/lib/src/types/query_types.dart
@@ -1,11 +1,22 @@
 import 'types.dart';
 
 class QueryOptions<T> {
+  /// Function that performs the query and returns the data as a `Future<T>`.
   final Future<T> Function() queryFn;
+
+  /// Key uniquely identifying this query.
   final List<Object> queryKey;
+
+  /// Staleness duration (in milliseconds) for the cached data.
   final double? staleTime;
+
+  /// Whether the query is enabled.
   final bool? enabled;
+
+  /// Whether the query should refetch on app restart.
   final bool? refetchOnRestart;
+
+  /// Whether the query should refetch on reconnect.
   final bool? refetchOnReconnect;
 
   const QueryOptions({
@@ -36,21 +47,27 @@ class QueryOptions<T> {
   }
 }
 
+/// Represents the current state of a query, including `status`, optional
+/// `data`, `error` and whether a fetch is ongoing.
 class QueryResult<T> {
   String key;
   QueryStatus status;
   T? data;
+
   /// Whether the query is currently fetching (background refetches etc).
   bool isFetching;
   Object? error;
 
-  QueryResult(this.key, this.status, this.data, this.error, {this.isFetching = false});
+  QueryResult(this.key, this.status, this.data, this.error,
+      {this.isFetching = false});
 
   bool get isError => status == QueryStatus.error;
   bool get isSuccess => status == QueryStatus.success;
   bool get isPending => status == QueryStatus.pending;
 }
 
+/// Result type returned by [useInfiniteQuery], with helper `fetchNextPage` and
+/// `isFetchingNextPage` flag.
 class InfiniteQueryResult<T> extends QueryResult<List<T>> {
   bool isFetchingNextPage;
   Function? fetchNextPage;