+++ packages/tanstack_query/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2 (2025-12-30)
+
+- Add dartdoc comments.
+
+## 1.0.1 (2025-12-28)
+
+- Add `useQueryClient` hook to retrieve the active `QueryClient` from widget context.
+
 ## 1.0.0 (18/12/2025)
 
 - First version with primary hooks (useQuery, useMutation, useInfiniteQuery)